requirements.txt
img2pdf==0.4.4
Pillow==9.4.0
tqdm==4.66.1